+++ src/Form/Extension/HelpMessageExtension.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Netzmacht\ContaoFormBundle\Form\Extension;
+
+use Symfony\Component\Form\AbstractTypeExtension;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\Form\FormView;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+final class HelpMessageExtension extends AbstractTypeExtension
+{
+    public function getExtendedType() : string
+    {
+        return FormType::class;
+    }
+
+    public function getExtendedTypes() : iterable
+    {
+        return [FormType::class];
+    }
+
+    public function buildView(FormView $view, FormInterface $form, array $options)
+    {
+        $view->vars['help'] = $options['help'] ?? '';
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults(['help' => null]);
+    }
+}